TINY-6800: moved test and add `editor.off` at the end of it
